+++ logfire/_internal/main.py
@@ -6,6 +6,7 @@ import json
 import sys
 import warnings
 from contextvars import Token
+from enum import Enum
 from functools import cached_property
 from time import time
 from typing import (
@@ -899,12 +900,17 @@ class Logfire:
     def _warn_if_not_initialized_for_instrumentation(self):
         self.config.warn_if_not_initialized('Instrumentation will have no effect')
 
-    def instrument_mcp(self) -> None:
-        """Instrument [MCP](https://modelcontextprotocol.io/) requests such as tool calls."""
+    def instrument_mcp(self, *, propagate_otel_context: bool = True) -> None:
+        """Instrument [MCP](https://modelcontextprotocol.io/) requests such as tool calls.
+
+        Args:
+            propagate_otel_context: Whether to enable propagation of the OpenTelemetry context.
+                Set to False to prevent setting extra fields like `traceparent` on the metadata of requests.
+        """
         from .integrations.mcp import instrument_mcp
 
         self._warn_if_not_initialized_for_instrumentation()
-        instrument_mcp(self)
+        instrument_mcp(self, propagate_otel_context)
 
     def instrument_pydantic(
         self,
@@ -962,6 +968,8 @@ class Logfire:
         /,
         *,
         event_mode: Literal['attributes', 'logs'] = 'attributes',
+        include_binary_content: bool | None = None,
+        **kwargs: Any,
     ) -> None: ...
 
     @overload
@@ -971,6 +979,8 @@ class Logfire:
         /,
         *,
         event_mode: Literal['attributes', 'logs'] = 'attributes',
+        include_binary_content: bool | None = None,
+        **kwargs: Any,
     ) -> pydantic_ai.models.Model: ...
 
     def instrument_pydantic_ai(
@@ -979,6 +989,8 @@ class Logfire:
         /,
         *,
         event_mode: Literal['attributes', 'logs'] | None = None,
+        include_binary_content: bool | None = None,
+        **kwargs: Any,
     ) -> pydantic_ai.models.Model | None:
         """Instrument PydanticAI.
 
@@ -989,14 +1001,24 @@ class Logfire:
                 If you pass a model, a new instrumented model will be returned.
             event_mode: See the [PydanticAI docs](https://ai.pydantic.dev/logfire/#data-format).
                 The default is whatever the default is in your version of PydanticAI.
+            include_binary_content: Whether to include base64 encoded binary content (e.g. images) in the events.
+                On by default. Requires PydanticAI 0.2.5 or newer.
+            kwargs: Additional keyword arguments to pass to
+                [`InstrumentationSettings`](https://ai.pydantic.dev/api/models/instrumented/#pydantic_ai.models.instrumented.InstrumentationSettings)
+                for future compatibility.
         """
         from .integrations.pydantic_ai import instrument_pydantic_ai
 
         self._warn_if_not_initialized_for_instrumentation()
+
+        if include_binary_content is not None:
+            kwargs['include_binary_content'] = include_binary_content
+
         return instrument_pydantic_ai(
             self,
             obj=obj,
             event_mode=event_mode,
+            **kwargs,
         )
 
     def instrument_fastapi(
@@ -2014,6 +2036,8 @@ class Logfire:
     ) -> None:
         """Create a counter metric that uses a callback to collect observations.
 
+        The callback is called every 60 seconds in a background thread.
+
         The counter metric is a cumulative metric that represents a single numerical value that only ever goes up.
 
         ```py
@@ -2056,6 +2080,8 @@ class Logfire:
     ) -> None:
         """Create a gauge metric that uses a callback to collect observations.
 
+        The callback is called every 60 seconds in a background thread.
+
         The gauge metric is a metric that represents a single numerical value that can arbitrarily go up and down.
 
         ```py
@@ -2095,6 +2121,8 @@ class Logfire:
         self, name: str, callbacks: Sequence[CallbackT], *, unit: str = '', description: str = ''
     ) -> None:
         """Create an up-down counter metric that uses a callback to collect observations.
+
+        The callback is called every 60 seconds in a background thread.
 
         The up-down counter is a cumulative metric that represents a single numerical value that can be adjusted up or
         down.
@@ -2428,7 +2456,9 @@ def prepare_otlp_attributes(attributes: dict[str, Any]) -> dict[str, otel_types.
 
 def prepare_otlp_attribute(value: Any) -> otel_types.AttributeValue:
     """Convert a user attribute to an OpenTelemetry compatible type."""
-    if isinstance(value, int):
+    if isinstance(value, Enum):
+        return logfire_json_dumps(value)
+    elif isinstance(value, int):
         if value > OTLP_MAX_INT_SIZE:
             warnings.warn(
                 f'Integer value {value} is larger than the maximum OTLP integer size of {OTLP_MAX_INT_SIZE} (64-bits), '